Basic JSHnit support and livereload watch task added.
Added a JSHint task to grunt, and updated watch tasks to cause JSHint to
be re-run every time a js file is saved in the directory.  And, added
that "open" flag to the server command, so the site auto-loads in your
default web browser.

Finally, this updates the the default version of dojo to the latest
patch release (1.9.2).